Add files via upload
Shopping Cart

Created in late 2021; This Java program simulates a basic shopping cart and checkout system for a fictional store named "MallWart." It starts by importing essential libraries and declaring variables to store customer information, item prices, quantities, costs, and more. Users are prompted to enter their name, email, and select quantities of five different items. The program calculates the subtotal, tax (13% of the subtotal), and the total cost of the purchased items. It then displays a receipt with a breakdown of the items, quantities, costs, and the final total. Additionally, it provides a message thanking the customer for shopping at MallWart and offers a chance to win a toaster through a contact email. The code concludes by displaying terms and conditions regarding the contest. This program serves as a simple example of a text-based shopping experience and demonstrates basic input/output operations, variable usage, and formatting techniques.