Daniyal's Arcade
Daniyal's Arcade

Created in late 2021; This Java program serves as a text-based arcade featuring two games: a Number Guessing Game and Rock-Paper-Scissors. It begins by presenting a menu to the user, allowing them to choose between these two games or exit. In the Number Guessing Game, the user has five attempts to correctly guess a randomly generated number, with hints provided if their guess is too high or too low. In the Rock-Paper-Scissors Game, the user plays against the CPU in a best-of-three match, with random choices generated for both sides. The program tallies wins for the player and CPU and declares the overall winner after three rounds.